+++ ffupdater/src/main/java/de/marmaro/krt/ffupdater/app/impl/FirefoxKlar.kt
@@ -31,14 +31,13 @@ class FirefoxKlar(private val apiConsumer: ApiConsumer) : BaseAppDetail() {
             apiConsumer = apiConsumer,
             task = "project.mobile.focus.release.latest",
             apkArtifact = "public/app-klar-$abiString-release-unsigned.apk",
-            keyForVersion = "tag_name",
+            keyForVersion = "version|tag_name",
             keyForReleaseDate = "published_at"
         )
         val result = mozillaCiConsumer.updateCheck()
-        val version = Regex("""^v(.*)$""").find(result.version)!!.groups[1]!!.value
         return AvailableVersionResult(
             downloadUrl = result.url,
-            version = version,
+            version = result.version,
             publishDate = result.releaseDate,
             fileSizeBytes = null,
             fileHash = null